android: Add ability to check for RenderDoc layer and patch it in.
After selecting an application to launch on Android, inspect it to see if
it contains the RenderDoc layer.  If it does not, display a warning similar
to desktop.  When clicked, offer to patch the APK and reinstall, with the
warning that it doesn't work for all applications (or at all for GLES).
Also provides pointers to how to package the layer yourself.

The process works by using the host temp directory to pull the APK and
modify it, using tools from the Android SDK.  If the steps fail for any
reason, the log is populated and patching is halted.