android: Check application for RenderDoc layer, offer to patch it in
At Launch time, inspect the application to see if it contains the
RenderDoc layer.  If it does not, offer to patch the APK and reinstall
it, with the warning that it doesn't work for all applications.  Also
provides pointers to how to package the layer yourself (TODO on Wiki).

The process works by using the host temp directory to pull the APK and
modify it, using tools from the Android SDK.

If the process fails for any reason, the log is populated and the process
is halted.

This WIP implementation (which is going to change) has the following caveats:
- Assumes armeabi-v7a
- Requires properly renamed layer to be in the following directory:
    /tmp/lib/armeabi-v7a/libVkLayer_RenderDoc.so
- Working on Linux and macOS, untested on Windows.
    If it compiles, it certainly won't check dependencies